refactor: extract auth hooks from providers
