+++ src/hooks/useAuthorization.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { AuthorizationContext } from '@/providers/AuthorizationProvider';
+
+export function useAuthorization() {
+  return useContext(AuthorizationContext);
+}
