+++ src/providers/AuthProvider.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useEffect, useMemo, useState } from 'react';
+import { createContext, useEffect, useMemo, useState } from 'react';
 import { supabase } from '@/lib/dataClient';
 import type { Session, User } from '@supabase/supabase-js';
 
@@ -8,7 +8,7 @@ interface AuthState {
   loading: boolean;
 }
 
-const AuthContext = createContext<AuthState>({ session: null, user: null, loading: true });
+export const AuthContext = createContext<AuthState>({ session: null, user: null, loading: true });
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const [session, setSession] = useState<Session | null>(null);
@@ -40,8 +40,3 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   const value = useMemo(() => ({ session, user, loading }), [session, user, loading]);
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
 }
-
-export function useAuth() {
-  return useContext(AuthContext);
-}
-
